[#151] refactor: useDebounce 함수 위치 변경
